backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.41
asyncpg==0.30.0
supabase==2.15.2
python-multipart==0.0.20
pydantic-settings==2.9.1
python-dotenv==1.1.0
alembic==1.15.2
pgvector==0.3.6
httpx==0.28.1
